changes in design and component naming
